+++ src/pages/ProjectListPage/components/SearchBarSection/SearchBarSection.tsx
@@ -1,21 +1,37 @@
-import { useEffect, useRef } from "react"
+import { useEffect, useRef, useState } from "react"
 import { IoMdSearch } from "react-icons/io"
 
 import { Box, Button, Center, Icon, useMediaQuery } from "@chakra-ui/react"
 
 import CommonInput from "@components/Input/CommonInput"
 
-import { SearchListSectionProps } from "../SearchListSection/SearchListSection"
+interface SearchBarSectionProps {
+  search: string
+  onSubmit: (keyword: string) => void
+}
 
-const SearchBarSection = ({ search }: SearchListSectionProps) => {
+const SearchBarSection = ({ search, onSubmit }: SearchBarSectionProps) => {
   const [isLargerThan768] = useMediaQuery("(min-width: 768px)")
   const $ref = useRef<HTMLInputElement>(null)
+  const [inputValue, setInputValue] = useState<string>(search)
+
+  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault()
+    if (!inputValue) {
+      console.log("검색어를 입력하세요")
+    }
+
+    onSubmit(inputValue)
+  }
+
+  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setInputValue(event.target.value)
+  }
 
   useEffect(() => {
     $ref.current?.focus()
   }, [])
 
-  console.log(search)
   return (
     <Box
       height="12rem"
@@ -25,7 +41,7 @@ const SearchBarSection = ({ search }: SearchListSectionProps) => {
         left="50%"
         top="20rem"
         transform="translate(-50%,-50%)">
-        <form onSubmit={() => console.log("enter")}>
+        <form onSubmit={handleSubmit}>
           <Box
             backgroundColor="white"
             borderRadius="5rem"
@@ -41,11 +57,14 @@ const SearchBarSection = ({ search }: SearchListSectionProps) => {
               height="7rem"
               fontSize="2xl"
               border="0"
+              marginLeft="1rem"
+              value={inputValue}
+              onChange={handleChange}
             />
             <Button
               position="absolute"
               top="2.3rem"
-              right="0"
+              right="1rem"
               backgroundColor="transparent"
               _hover={{ backgroundColor: "transparent" }}>
               <Icon
